pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "activemirror"
version = "0.1.0"
description = "Intelligence that remembers - Persistent, reflective AI experiences"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "MirrorDNA-Reflection-Protocol", email = "contact@mirrordna.org"},
]
keywords = ["ai", "memory", "conversation", "persistence", "mirrordna", "lingos"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
    "black>=23.0",
]

postgresql = [
    "psycopg2-binary>=2.9",
]

all = [
    "activemirror[postgresql]",
]

[project.urls]
Homepage = "https://github.com/MirrorDNA-Reflection-Protocol/ActiveMirrorOS"
Documentation = "https://github.com/MirrorDNA-Reflection-Protocol/ActiveMirrorOS/tree/main/docs"
Repository = "https://github.com/MirrorDNA-Reflection-Protocol/ActiveMirrorOS"
Issues = "https://github.com/MirrorDNA-Reflection-Protocol/ActiveMirrorOS/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["activemirror*"]
namespaces = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[tool.ruff]
line-length = 88
target-version = "py38"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = []

[tool.ruff.isort]
known-first-party = ["activemirror"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
